--- frappe/core/doctype/customize_form/customize_form.json
@@ -1,123 +1,0 @@
-{
- "autoname": "DL.####", 
- "creation": "2013-01-29 17:55:08", 
- "docstatus": 0, 
- "doctype": "DocType", 
- "fields": [
-  {
-   "fieldname": "doc_type", 
-   "fieldtype": "Link", 
-   "hidden": 0, 
-   "in_list_view": 1, 
-   "label": "Enter Form Type", 
-   "no_copy": 0, 
-   "options": "DocType", 
-   "permlevel": 0, 
-   "search_index": 0
-  }, 
-  {
-   "fieldname": "properties", 
-   "fieldtype": "Section Break", 
-   "label": "Properties", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "column_break0", 
-   "fieldtype": "Column Break", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "default_print_format", 
-   "fieldtype": "Link", 
-   "in_list_view": 1, 
-   "label": "Default Print Format", 
-   "no_copy": 0, 
-   "options": "Print Format", 
-   "permlevel": 0, 
-   "search_index": 0
-  }, 
-  {
-   "description": "Fields separated by comma (,) will be included in the<br /><b>Search By</b> list of Search dialog box", 
-   "fieldname": "search_fields", 
-   "fieldtype": "Data", 
-   "in_list_view": 1, 
-   "label": "Search Fields", 
-   "no_copy": 0, 
-   "permlevel": 0, 
-   "search_index": 0
-  }, 
-  {
-   "fieldname": "sort_field", 
-   "fieldtype": "Data", 
-   "label": "Sort Field", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "sort_order", 
-   "fieldtype": "Data", 
-   "label": "Sort Order", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "column_break1", 
-   "fieldtype": "Column Break", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "allow_copy", 
-   "fieldtype": "Check", 
-   "label": "Hide Copy", 
-   "no_copy": 0, 
-   "permlevel": 0, 
-   "search_index": 0
-  }, 
-  {
-   "depends_on": "", 
-   "fieldname": "max_attachments", 
-   "fieldtype": "Int", 
-   "label": "Max Attachments", 
-   "no_copy": 0, 
-   "permlevel": 0, 
-   "search_index": 0
-  }, 
-  {
-   "description": "Customize Label, Print Hide, Default etc.", 
-   "fieldname": "fields_section_break", 
-   "fieldtype": "Section Break", 
-   "label": "Fields", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "customize_form_fields", 
-   "fieldtype": "Table", 
-   "label": "Fields", 
-   "no_copy": 0, 
-   "options": "Customize Form Field", 
-   "permlevel": 0, 
-   "search_index": 0
-  }
- ], 
- "hide_toolbar": 1, 
- "icon": "icon-glass", 
- "idx": 1, 
- "issingle": 1, 
- "modified": "2014-08-22 05:42:45.083260", 
- "modified_by": "Administrator", 
- "module": "Core", 
- "name": "Customize Form", 
- "owner": "Administrator", 
- "permissions": [
-  {
-   "create": 1, 
-   "email": 1, 
-   "permlevel": 0, 
-   "print": 1, 
-   "read": 1, 
-   "report": 0, 
-   "role": "System Manager", 
-   "submit": 0, 
-   "write": 1
-  }
- ], 
- "search_fields": "doc_type"
-}
