+++ frappe/core/doctype/doctype/doctype.py
@@ -10,7 +10,7 @@ from frappe.utils import now, cint
 from frappe.model import no_value_fields
 from frappe.model.document import Document
 from frappe.model.db_schema import type_map
-from frappe.core.doctype.property_setter.property_setter import make_property_setter
+from frappe.custom.doctype.property_setter.property_setter import make_property_setter
 from frappe.core.doctype.notification_count.notification_count import delete_notification_count_for
 from frappe.modules import make_boilerplate
