--- frappe/core/doctype/property_setter/property_setter.json
@@ -1,122 +1,0 @@
-{
- "creation": "2013-01-10 16:34:04", 
- "description": "Property Setter overrides a standard DocType or Field property", 
- "docstatus": 0, 
- "doctype": "DocType", 
- "fields": [
-  {
-   "fieldname": "help", 
-   "fieldtype": "HTML", 
-   "label": "Help", 
-   "options": "<div class=\"alert\">Please don't update it as it can mess up your form. Use the Customize Form View and Custom Fields to set properties!</div>", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "sb0", 
-   "fieldtype": "Section Break", 
-   "permlevel": 0
-  }, 
-  {
-   "depends_on": "eval:doc.__islocal", 
-   "fieldname": "doctype_or_field", 
-   "fieldtype": "Select", 
-   "in_list_view": 1, 
-   "label": "DocType or Field", 
-   "options": "\nDocField\nDocType", 
-   "permlevel": 0, 
-   "reqd": 1
-  }, 
-  {
-   "description": "New value to be set", 
-   "fieldname": "value", 
-   "fieldtype": "Text", 
-   "in_list_view": 1, 
-   "label": "Set Value", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "column_break0", 
-   "fieldtype": "Column Break", 
-   "permlevel": 0, 
-   "search_index": 0
-  }, 
-  {
-   "fieldname": "doc_type", 
-   "fieldtype": "Link", 
-   "in_filter": 0, 
-   "label": "DocType", 
-   "options": "DocType", 
-   "permlevel": 0, 
-   "reqd": 1, 
-   "search_index": 1
-  }, 
-  {
-   "depends_on": "eval:doc.doctype_or_field=='DocField'", 
-   "description": "ID (name) of the entity whose property is to be set", 
-   "fieldname": "field_name", 
-   "fieldtype": "Data", 
-   "in_filter": 0, 
-   "label": "Field Name", 
-   "permlevel": 0, 
-   "reqd": 0, 
-   "search_index": 1
-  }, 
-  {
-   "fieldname": "property", 
-   "fieldtype": "Data", 
-   "in_filter": 0, 
-   "label": "Property", 
-   "permlevel": 0, 
-   "reqd": 1, 
-   "search_index": 1
-  }, 
-  {
-   "fieldname": "property_type", 
-   "fieldtype": "Data", 
-   "label": "Property Type", 
-   "permlevel": 0
-  }, 
-  {
-   "fieldname": "default_value", 
-   "fieldtype": "Data", 
-   "label": "Default Value", 
-   "permlevel": 0
-  }
- ], 
- "icon": "icon-glass", 
- "idx": 1, 
- "modified": "2014-07-24 02:04:26.838056", 
- "modified_by": "Administrator", 
- "module": "Core", 
- "name": "Property Setter", 
- "owner": "Administrator", 
- "permissions": [
-  {
-   "cancel": 0, 
-   "create": 1, 
-   "delete": 1, 
-   "email": 1, 
-   "permlevel": 0, 
-   "print": 1, 
-   "read": 1, 
-   "report": 1, 
-   "role": "Administrator", 
-   "submit": 0, 
-   "write": 1
-  }, 
-  {
-   "cancel": 0, 
-   "create": 1, 
-   "delete": 1, 
-   "email": 1, 
-   "permlevel": 0, 
-   "print": 1, 
-   "read": 1, 
-   "report": 1, 
-   "role": "System Manager", 
-   "submit": 0, 
-   "write": 1
-  }
- ], 
- "search_fields": "doc_type,property"
-}
