+++ frappe/custom/doctype/custom_field/custom_field.json
@@ -1,0 +1,304 @@
+{
+ "creation": "2013-01-10 16:34:01",
+ "description": "Adds a custom field to a DocType",
+ "docstatus": 0,
+ "doctype": "DocType",
+ "fields": [
+  {
+   "fieldname": "dt",
+   "fieldtype": "Link",
+   "in_filter": 1,
+   "in_list_view": 1,
+   "label": "Document",
+   "no_copy": 0,
+   "oldfieldname": "dt",
+   "oldfieldtype": "Link",
+   "options": "DocType",
+   "permlevel": 0,
+   "reqd": 1,
+   "search_index": 1
+  },
+  {
+   "fieldname": "label",
+   "fieldtype": "Data",
+   "in_filter": 1,
+   "label": "Label",
+   "no_copy": 1,
+   "oldfieldname": "label",
+   "oldfieldtype": "Data",
+   "permlevel": 0,
+   "reqd": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "label_help",
+   "fieldtype": "HTML",
+   "label": "Label Help",
+   "oldfieldtype": "HTML",
+   "permlevel": 0
+  },
+  {
+   "description": "Select the label after which you want to insert new field.",
+   "fieldname": "insert_after",
+   "fieldtype": "Select",
+   "label": "Insert After",
+   "no_copy": 1,
+   "oldfieldname": "insert_after",
+   "oldfieldtype": "Select",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "fieldtype",
+   "fieldtype": "Select",
+   "in_filter": 1,
+   "in_list_view": 1,
+   "label": "Field Type",
+   "no_copy": 0,
+   "oldfieldname": "fieldtype",
+   "oldfieldtype": "Select",
+   "options": "Button\nCheck\nCode\nColumn Break\nCurrency\nData\nDate\nDatetime\nDynamic Link\nFloat\nHTML\nImage\nInt\nLink\nLong Text\nPassword\nPercent\nRead Only\nSection Break\nSelect\nSmall Text\nTable\nText\nText Editor\nTime",
+   "permlevel": 0,
+   "reqd": 1,
+   "search_index": 0
+  },
+  {
+   "depends_on": "eval:in_list([\"Float\", \"Currency\", \"Percent\"], doc.fieldtype)",
+   "description": "Set non-standard precision for a Float or Currency field",
+   "fieldname": "precision",
+   "fieldtype": "Select",
+   "label": "Precision",
+   "options": "\n1\n2\n3\n4\n5\n6",
+   "permlevel": 0,
+   "precision": ""
+  },
+  {
+   "fieldname": "options_help",
+   "fieldtype": "HTML",
+   "label": "Options Help",
+   "oldfieldtype": "HTML",
+   "permlevel": 0
+  },
+  {
+   "fieldname": "options",
+   "fieldtype": "Text",
+   "in_list_view": 1,
+   "label": "Options",
+   "no_copy": 0,
+   "oldfieldname": "options",
+   "oldfieldtype": "Text",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "description",
+   "fieldtype": "Text",
+   "label": "Field Description",
+   "no_copy": 0,
+   "oldfieldname": "description",
+   "oldfieldtype": "Text",
+   "permlevel": 0,
+   "print_width": "300px",
+   "search_index": 0,
+   "width": "300px"
+  },
+  {
+   "fieldname": "fieldname",
+   "fieldtype": "Data",
+   "in_list_view": 1,
+   "label": "Fieldname",
+   "no_copy": 1,
+   "oldfieldname": "fieldname",
+   "oldfieldtype": "Data",
+   "permlevel": 0,
+   "read_only": 1,
+   "search_index": 0
+  },
+  {
+   "fieldname": "properties",
+   "fieldtype": "Column Break",
+   "label": "Properties",
+   "oldfieldtype": "Column Break",
+   "permlevel": 0,
+   "print_width": "50%",
+   "width": "50%"
+  },
+  {
+   "default": "0",
+   "fieldname": "permlevel",
+   "fieldtype": "Int",
+   "label": "Permission Level",
+   "no_copy": 0,
+   "oldfieldname": "permlevel",
+   "oldfieldtype": "Int",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "reqd",
+   "fieldtype": "Check",
+   "in_list_view": 1,
+   "label": "Is Mandatory Field",
+   "no_copy": 0,
+   "oldfieldname": "reqd",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "read_only",
+   "fieldtype": "Check",
+   "label": "Read Only",
+   "permlevel": 0
+  },
+  {
+   "depends_on": "eval:doc.fieldtype===\"Link\"",
+   "fieldname": "ignore_user_permissions",
+   "fieldtype": "Check",
+   "label": "Ignore User Permissions",
+   "permlevel": 0
+  },
+  {
+   "fieldname": "width",
+   "fieldtype": "Data",
+   "label": "Width",
+   "no_copy": 0,
+   "oldfieldname": "width",
+   "oldfieldtype": "Data",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "default",
+   "fieldtype": "Text",
+   "label": "Default Value",
+   "no_copy": 0,
+   "oldfieldname": "default",
+   "oldfieldtype": "Text",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "hidden",
+   "fieldtype": "Check",
+   "label": "Hidden",
+   "permlevel": 0
+  },
+  {
+   "fieldname": "depends_on",
+   "fieldtype": "Data",
+   "label": "Depends On",
+   "permlevel": 0
+  },
+  {
+   "fieldname": "print_hide",
+   "fieldtype": "Check",
+   "label": "Print Hide",
+   "no_copy": 0,
+   "oldfieldname": "print_hide",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "print_width",
+   "fieldtype": "Data",
+   "hidden": 1,
+   "label": "Print Width",
+   "no_copy": 1,
+   "permlevel": 0,
+   "print_hide": 1
+  },
+  {
+   "fieldname": "no_copy",
+   "fieldtype": "Check",
+   "label": "No Copy",
+   "no_copy": 0,
+   "oldfieldname": "no_copy",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "allow_on_submit",
+   "fieldtype": "Check",
+   "label": "Allow on Submit",
+   "no_copy": 0,
+   "oldfieldname": "allow_on_submit",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "in_filter",
+   "fieldtype": "Check",
+   "label": "In Report Filter",
+   "no_copy": 0,
+   "oldfieldname": "in_filter",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "in_list_view",
+   "fieldtype": "Check",
+   "label": "In List View",
+   "permlevel": 0
+  },
+  {
+   "fieldname": "report_hide",
+   "fieldtype": "Check",
+   "label": "Report Hide",
+   "no_copy": 0,
+   "oldfieldname": "report_hide",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "search_index",
+   "fieldtype": "Check",
+   "hidden": 1,
+   "label": "Index",
+   "no_copy": 1,
+   "permlevel": 0,
+   "print_hide": 1
+  }
+ ],
+ "icon": "icon-glass",
+ "idx": 1,
+ "modified": "2014-09-05 07:41:13.076821",
+ "modified_by": "Administrator",
+ "module": "Custom",
+ "name": "Custom Field",
+ "owner": "Administrator",
+ "permissions": [
+  {
+   "cancel": 0,
+   "create": 1,
+   "delete": 1,
+   "email": 1,
+   "permlevel": 0,
+   "print": 1,
+   "read": 1,
+   "report": 1,
+   "role": "Administrator",
+   "submit": 0,
+   "write": 1
+  },
+  {
+   "cancel": 0,
+   "create": 1,
+   "delete": 1,
+   "email": 1,
+   "permlevel": 0,
+   "print": 1,
+   "read": 1,
+   "report": 1,
+   "role": "System Manager",
+   "submit": 0,
+   "write": 1
+  }
+ ],
+ "search_fields": "dt,label,fieldtype,options"
+}
