+++ frappe/custom/doctype/customize_form_field/customize_form_field.json
@@ -1,0 +1,288 @@
+{
+ "allow_copy": 0,
+ "autoname": "DLF.#####",
+ "creation": "2013-02-22 01:27:32",
+ "docstatus": 0,
+ "doctype": "DocType",
+ "fields": [
+  {
+   "fieldname": "label_and_type",
+   "fieldtype": "Section Break",
+   "label": "Label and Type",
+   "permlevel": 0,
+   "precision": ""
+  },
+  {
+   "fieldname": "label",
+   "fieldtype": "Data",
+   "hidden": 0,
+   "in_list_view": 1,
+   "label": "Label",
+   "oldfieldname": "label",
+   "oldfieldtype": "Data",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 0,
+   "search_index": 1
+  },
+  {
+   "default": "Data",
+   "fieldname": "fieldtype",
+   "fieldtype": "Select",
+   "hidden": 0,
+   "in_list_view": 1,
+   "label": "Type",
+   "oldfieldname": "fieldtype",
+   "oldfieldtype": "Select",
+   "options": "Attach\nButton\nCheck\nCode\nColumn Break\nCurrency\nData\nDate\nDatetime\nDynamic Link\nFloat\nFold\nHTML\nImage\nInt\nLink\nLong Text\nPassword\nPercent\nRead Only\nSection Break\nSelect\nSmall Text\nTable\nText\nText Editor\nTime",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 1,
+   "search_index": 1
+  },
+  {
+   "fieldname": "fieldname",
+   "fieldtype": "Data",
+   "hidden": 0,
+   "in_list_view": 1,
+   "label": "Name",
+   "oldfieldname": "fieldname",
+   "oldfieldtype": "Data",
+   "permlevel": 0,
+   "print_hide": 0,
+   "read_only": 1,
+   "reqd": 0,
+   "search_index": 1
+  },
+  {
+   "fieldname": "reqd",
+   "fieldtype": "Check",
+   "hidden": 0,
+   "label": "Mandatory",
+   "oldfieldname": "reqd",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "print_hide": 0,
+   "print_width": "50px",
+   "reqd": 0,
+   "search_index": 0,
+   "width": "50px"
+  },
+  {
+   "fieldname": "in_list_view",
+   "fieldtype": "Check",
+   "label": "In List View",
+   "permlevel": 0
+  },
+  {
+   "fieldname": "column_break_7",
+   "fieldtype": "Column Break",
+   "permlevel": 0,
+   "precision": ""
+  },
+  {
+   "depends_on": "eval:in_list([\"Float\", \"Currency\", \"Percent\"], doc.fieldtype)",
+   "description": "Set non-standard precision for a Float or Currency field",
+   "fieldname": "precision",
+   "fieldtype": "Select",
+   "label": "Precision",
+   "options": "\n1\n2\n3\n4\n5\n6",
+   "permlevel": 0,
+   "precision": ""
+  },
+  {
+   "description": "For Links, enter the DocType as range\nFor Select, enter list of Options separated by comma",
+   "fieldname": "options",
+   "fieldtype": "Text",
+   "hidden": 0,
+   "in_list_view": 1,
+   "label": "Options",
+   "oldfieldname": "options",
+   "oldfieldtype": "Text",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "permissions",
+   "fieldtype": "Section Break",
+   "label": "Permissions",
+   "permlevel": 0,
+   "precision": ""
+  },
+  {
+   "description": "This field will appear only if the fieldname defined here has value OR the rules are true (examples): <br>\nmyfield\neval:doc.myfield=='My Value'<br>\neval:doc.age>18",
+   "fieldname": "depends_on",
+   "fieldtype": "Data",
+   "hidden": 0,
+   "label": "Depends On",
+   "oldfieldname": "depends_on",
+   "oldfieldtype": "Data",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 0
+  },
+  {
+   "default": "0",
+   "fieldname": "permlevel",
+   "fieldtype": "Int",
+   "hidden": 0,
+   "in_list_view": 1,
+   "label": "Perm Level",
+   "oldfieldname": "permlevel",
+   "oldfieldtype": "Int",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "hidden",
+   "fieldtype": "Check",
+   "hidden": 0,
+   "label": "Hidden",
+   "oldfieldname": "hidden",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "print_hide": 0,
+   "print_width": "50px",
+   "reqd": 0,
+   "search_index": 0,
+   "width": "50px"
+  },
+  {
+   "fieldname": "column_break_14",
+   "fieldtype": "Column Break",
+   "permlevel": 0,
+   "precision": ""
+  },
+  {
+   "fieldname": "ignore_user_permissions",
+   "fieldtype": "Check",
+   "label": "Ignore User Permissions",
+   "permlevel": 0
+  },
+  {
+   "fieldname": "allow_on_submit",
+   "fieldtype": "Check",
+   "hidden": 0,
+   "label": "Allow on Submit",
+   "oldfieldname": "allow_on_submit",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 0
+  },
+  {
+   "fieldname": "report_hide",
+   "fieldtype": "Check",
+   "hidden": 0,
+   "label": "Report Hide",
+   "oldfieldname": "report_hide",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 0
+  },
+  {
+   "fieldname": "display",
+   "fieldtype": "Section Break",
+   "label": "Display",
+   "permlevel": 0,
+   "precision": ""
+  },
+  {
+   "fieldname": "default",
+   "fieldtype": "Text",
+   "hidden": 0,
+   "label": "Default",
+   "oldfieldname": "default",
+   "oldfieldtype": "Text",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 0,
+   "search_index": 0
+  },
+  {
+   "fieldname": "in_filter",
+   "fieldtype": "Check",
+   "hidden": 0,
+   "label": "In Filter",
+   "oldfieldname": "in_filter",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "print_hide": 0,
+   "print_width": "50px",
+   "reqd": 0,
+   "width": "50px"
+  },
+  {
+   "fieldname": "column_break_21",
+   "fieldtype": "Column Break",
+   "permlevel": 0,
+   "precision": ""
+  },
+  {
+   "fieldname": "description",
+   "fieldtype": "Text",
+   "hidden": 0,
+   "label": "Description",
+   "oldfieldname": "description",
+   "oldfieldtype": "Text",
+   "permlevel": 0,
+   "print_hide": 0,
+   "print_width": "300px",
+   "reqd": 0,
+   "width": "300px"
+  },
+  {
+   "fieldname": "print_hide",
+   "fieldtype": "Check",
+   "hidden": 0,
+   "label": "Print Hide",
+   "oldfieldname": "print_hide",
+   "oldfieldtype": "Check",
+   "permlevel": 0,
+   "print_hide": 0,
+   "reqd": 0,
+   "search_index": 0
+  },
+  {
+   "description": "Print Width of the field, if the field is a column in a table",
+   "fieldname": "print_width",
+   "fieldtype": "Data",
+   "label": "Print Width",
+   "permlevel": 0,
+   "print_width": "50px",
+   "width": "50px"
+  },
+  {
+   "fieldname": "width",
+   "fieldtype": "Data",
+   "hidden": 0,
+   "in_list_view": 1,
+   "label": "Width",
+   "oldfieldname": "width",
+   "oldfieldtype": "Data",
+   "permlevel": 0,
+   "print_hide": 0,
+   "print_width": "50px",
+   "reqd": 0,
+   "search_index": 0,
+   "width": "50px"
+  }
+ ],
+ "hide_heading": 0,
+ "hide_toolbar": 0,
+ "idx": 1,
+ "issingle": 0,
+ "istable": 1,
+ "modified": "2014-09-05 07:41:29.641455",
+ "modified_by": "Administrator",
+ "module": "Custom",
+ "name": "Customize Form Field",
+ "owner": "Administrator",
+ "permissions": [],
+ "read_only": 0
+}
