+++ frappe/custom/doctype/property_setter/property_setter.js
