+++ frappe/modules.txt
@@ -2,3 +2,4 @@ Core
 Website
 Workflow
 Email
+Custom
